+++ wire-ios/Wire-iOS/Sources/UserInterface/DeviceView/Views/CertificateDetailsView.swift
