+++ wire-ios/Wire-iOS/Sources/UserInterface/DeviceView/Views/CopyValueView.swift
