+++ wire-ios/Wire-iOS/Sources/UserInterface/DeviceView/Views/DeviceDetailsE2EIdentityCertificateView.swift
