+++ wire-ios/Wire-iOS/Sources/UserInterface/DeviceView/Views/DeviceView.swift
